Add OCPP 1.6J client, smart charging, auth list and offline journal
Implements the Core profile in full plus SmartCharging,
LocalAuthListManagement, Reservation and RemoteTrigger. One task owns the
WebSocket and all OCPP state; callers enqueue rather than blocking on the
network, and outstanding CALLs time out so an unanswered Authorize cannot
leave a driver waiting at the connector.

Behaviour worth calling out:
  - A remote start still goes through normal authorisation. The CSMS asking
    is not the same as the CSMS having checked.
  - A Soft reset stops the transaction cleanly first; only Hard resets cut
    a charging session.
  - Taking a connector out of service mid-session answers "Scheduled", not
    "Accepted".
  - An unsupported action gets a CALLERROR, since silence makes some CSMS
    implementations tear down the whole session.
  - Reconnection backs off exponentially to 5 minutes, so a backend outage
    does not become a thundering herd on recovery.

The transaction journal writes to flash before transmitting and replays with
original timestamps, since re-stamping on reconnect moves sessions to
whenever the link returned. Acknowledgement is an appended record rather
than an edited flag, because flash cannot be rewritten in place.

Test infrastructure: five host suites plus a compile check that builds every
App/ source against HAL and FreeRTOS stubs. The check found a struct that
was not the size its static assertion claimed, two missing includes and an
undeclared peripheral symbol.

+++ firmware/App/Src/evse_config.c
@@ -7,6 +7,7 @@
 #include "FreeRTOS.h"
 #include "task.h"
 #include <stddef.h>
+#include <stdint.h>
 #include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -129,9 +130,11 @@ void cfg_load_defaults(evse_config_t *cfg)
 /** Which physical copy we most recently loaded from, so we write the other. */
 static uint32_t s_active_addr;
 
+/* uintptr_t rather than a direct cast so this is also clean when built for the
+ * host compile-check, where pointers are wider than the flash address. */
 static const evse_config_t *cfg_at(uint32_t addr)
 {
-    return (const evse_config_t *)addr;
+    return (const evse_config_t *)(uintptr_t)addr;
 }
 
 static bool flash_erase_sector(uint32_t addr)
@@ -172,7 +175,7 @@ static bool flash_write_config(uint32_t addr, const evse_config_t *c)
     }
     HAL_FLASH_Lock();
 
-    return ok && memcmp((const void *)addr, c, sizeof(*c)) == 0;
+    return ok && memcmp((const void *)(uintptr_t)addr, c, sizeof(*c)) == 0;
 }
 
 void cfg_init(void)
